+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### 🚀 Updates
 
 - Added support for installing the canary release (when applicable).
+- Brought back support for detecting a version from `package.json` engines.
 - Updated to support proto v0.17 release.
 
 ## 0.2.1